requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
transformers>=4.30.0

# Medical Imaging
monai>=1.2.0
SimpleITK>=2.2.0
pydicom>=2.3.0

# Computer Vision
opencv-python>=4.7.0
albumentations>=1.3.0
scikit-image>=0.20.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
pillow>=9.5.0

# Machine Learning
scikit-learn>=1.2.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Experiment Tracking (Optional)
tensorboard>=2.13.0
wandb>=0.15.0

# Development
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0

# Documentation
sphinx>=6.2.0
